_requirements.txt
frictionless==4.1.0
python-dotenv==0.15.0